backend/requirements.txt
asgiref==3.11.0
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.14.0
dotenv==0.9.9
implicit==0.7.2
joblib==1.5.2
numpy==2.3.5
psycopg2-binary==2.9.10
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
sqlparse==0.5.4
threadpoolctl==3.6.0
tqdm==4.67.1
